+++ src-tauri/src/commands/mod.rs
@@ -4,3 +4,8 @@ pub mod mastery_tracker {
     pub mod get_all_user_progress;
     pub mod save_item_progress;
 }
+
+pub mod weekly_tracker {
+    pub mod get_weekly_tasks;
+    pub mod adjust_weekly_task;
+}