+++ src/pages/WeeklyTracker/index.ts
@@ -1,0 +1,1 @@
+export { WeeklyTracker } from "./WeeklyTracker";